transformer_project/pyproject.toml
[project]
name = "transformer_project"
version = "0.1.0"
description = "Transformer implementation for German-English translation"
authors = [
    {name = "Abtin", email = "abtin.pourhadi@hhu.de"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "torch (>=2.10.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "datasets (>=4.5.0,<5.0.0)",
    "tokenizers (>=0.22.2,<0.23.0)",
    "transformers (>=4.57.6,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "wandb (>=0.24.0,<0.25.0)",
    "sacrebleu (>=2.6.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "tabulate (>=0.9.0,<1.0.0)"
]

[tool.poetry]
packages = [
    {include = "modelling"},
    {include = "text"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
